Forward-declare `best::span` (#28)
With a little bit of cleverness, we can add `best::span` to the `fwd.h`
header.

The primary obstacle was that it has a `best::option<size_t>` template
parameter, which is not defined until way later. However, by making that
`size_t` dependent on the first template parameter, we eliminate this
problem by forcing two-phase lookup.